QoL improvements across DashboardActivity, NotificationsActivity, and Utils.
