+++ Phone App/android_app/BiofeedbackStress/build.gradle
@@ -7,8 +7,8 @@ android {
         applicationId "com.example.biofeedbackstress"
         minSdkVersion 21
         targetSdkVersion 29
-        versionCode 9
-        versionName "1.81"
+        versionCode 10
+        versionName "1.8.2"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
